Refactor performance bench code
